requirements.txt
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
psycopg2-binary==2.9.1
pytz==2021.1
Serializer==0.2.1
simplejson==3.17.5
sqlparse==0.4.1
